+++ archives/contracts-dataset-parser.php
@@ -5,7 +5,7 @@
 
 // toobs2017@gmail.com and the GoC-Spending team!
 
-require('contracts-helpers.php');
+require('Helpers.php');
 
 // Go crazy!
 ini_set('memory_limit', '2048M');
